refactor(standards): fix coding standard violations across codebase (Wave 1)
- Replace raw int/float with i32/f32 in event structs (MouseEvents, InputEvents, ScreenEvents)
- Replace const char* with std::string_view in AnimationEvents
- Add explicit casts at Raylib boundaries in InputSystem
- Fix App::Config to use i32 types and designated initializers in main.cpp
- Add [[nodiscard]] to Application::run(), EventManager::instance(), FontManager::instance()
- Add noexcept to Renderer::screenWidth()/screenHeight()
- Change local constexpr to static constexpr in PausePopupScreen.cpp and MainMenuScreen.cpp
- Extract duplicated unload logic in ShaderManager to private helper unloadExisting()

+++ src/Systems/Input/InputSystem.cpp
@@ -20,16 +20,16 @@ void InputSystem::poll() {
     const Vector2 mouse = GetMousePosition();
     for (const i32 btn : {MOUSE_BUTTON_LEFT, MOUSE_BUTTON_RIGHT, MOUSE_BUTTON_MIDDLE}) {
         if (IsMouseButtonPressed(btn)) {
-            bus.trigger(event::mouse::MousePressedEvent{btn, mouse.x, mouse.y});
+            bus.trigger(event::mouse::MousePressedEvent{btn, static_cast<f32>(mouse.x), static_cast<f32>(mouse.y)});
         }
         if (IsMouseButtonReleased(btn)) {
-            bus.trigger(event::mouse::MouseReleasedEvent{btn, mouse.x, mouse.y});
+            bus.trigger(event::mouse::MouseReleasedEvent{btn, static_cast<f32>(mouse.x), static_cast<f32>(mouse.y)});
         }
     }
 
     const Vector2 scroll = GetMouseWheelMoveV();
     if (scroll.x != 0.0f || scroll.y != 0.0f) {
-        bus.trigger(event::mouse::MouseScrolledEvent{scroll.x, scroll.y});
+        bus.trigger(event::mouse::MouseScrolledEvent{static_cast<f32>(scroll.x), static_cast<f32>(scroll.y)});
     }
 
     if (WindowShouldClose()) {
